Worked on Normal Distribution.
Normal Distribution is widely used because it follows a pattern to describe the results.
In a normal distribution:

Most of the data falls near the mean, making it the highest point on the curve.
As you move away from the mean, the frequency of data points decreases symmetrically.
The tails of the curve extend infinitely in both directions, theoretically.
Now, why is it so popular? Well, lots of things in nature tend to follow this pattern. For example, heights of people, IQ scores, errors in measurements, and so on. That's why it's often called the "bell curve" because, well, it looks like a bell!

One super cool thing about the normal distribution is the 68-95-99.7 rule (also known as the empirical rule). It says that: