learn how to test the model
Train and Test method is used to train and test the model.
From the data sets we divide it into training datasets and testing datasets.

First, we train the model with trained datasets.
After training the model we'll check the R-squared value (Remainder) of the model using trained datasets.
R squared value checks how well our data fits into the model.
Next, we go with testing datasets. We'll check the R-squared value of the model using test datasets.
if the R-squared values are good enough. Then we can go with the model to predict the values.