Wow Wow Wow
DrawBack :

It will not give same answer if you run multiple times. That means it doesn't provide 100% accurate answer.
Because the DecisionTree returns values based on the Probability.



Worked with DecisionTree.
Same as Regreesiopn Models.

Procedure:

First, We build the TreeModel using the DesicsionTreeClassifier method from sklearn.tree module.

Next, We train the model with model with datasets.(CSV files) using fit()

Now , the model is ready is to use.

Now go with predict method. By providing the available constraint attributes we ask model to predict the future values.